+++ .env-example
@@ -6,3 +6,6 @@ POSTGRES_PORT=5432
 POSTGRES_HOST=local_database
 
 DEBUG=true
+
+JWT_SECRET:53u4irihjiodfgyuingmdfs
+JWE_ALGORITHM=HS256
