+++ backend_api/app/settings.py
@@ -12,6 +12,9 @@ class Settings(BaseSettings):
 
     DEBUG: bool = False
 
+    JWT_SECRET:str
+    JWE_ALGORITHM:str
+
     @property
     def DATABASE_URL_ASYNC(self) -> str:
         return (
